refactor: hackathon module.css corrections and removed style.css
