+++ src/modules/Dashboard/modules/Hackathon/pages/HackathonCreate.module.css
@@ -539,3 +539,201 @@ input[type=file]::file-selector-button:hover {
         font-size: .8rem;
     }
 }
+
+.toolTipStyle {
+    background-color: var(--blue);
+    color: var(--White);
+    border-radius: var(--borderRadius);
+}
+
+.hackathonBox {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 1rem;
+}
+
+@media(width<=1270px) {
+    .hackathonBox {
+        grid-template-columns: 1fr 1fr;
+    }
+}
+
+.hackathonBox .cardComponent {
+    background-color: #ffffff;
+    border: 1px solid #dbdbdb;
+    border-radius: 10px;
+    display: flex;
+    flex-direction: column;
+    padding: 1.5rem;
+    text-align: left;
+}
+
+.hackathonBox .cardComponent:hover {
+    box-shadow: 0px 0px 10px .5px rgba(0, 0, 0, 0.1);
+}
+
+.hackathonBox .frame {
+    display: flex;
+    flex-direction: column;
+    gap: 1.3rem;
+    position: relative;
+    width: fit-content;
+}
+
+.hackathonBox .div {
+    align-items: flex-start;
+    display: flex;
+    gap: 85.5px;
+    position: relative;
+    width: fit-content;
+}
+
+.hackathonBox .title {
+    height: 35.81px;
+    min-width: 147.9px;
+    position: relative;
+}
+
+.hackathonBox .textWrapper {
+    color: var(--blue);
+    font-size: 22.6px;
+    font-weight: 700;
+    text-align: left;
+    letter-spacing: 0;
+    line-height: normal;
+    white-space: nowrap;
+    width: 144px;
+}
+
+.hackathonBox .textWrapper2 {
+    color: var(--black);
+    font-size: 12.9px;
+    font-weight: 300;
+    left: 0;
+    line-height: normal;
+    position: absolute;
+    top: 26px;
+    white-space: nowrap;
+}
+
+.hackathonBox .shared {
+    height: 24.52px;
+    min-width: 61.95px;
+    position: relative;
+}
+
+.hackathonBox .frame2 {
+    align-items: flex-start;
+    display: flex;
+    gap: 12px;
+    position: relative;
+    width: 62px;
+}
+
+.hackathonBox .group {
+    background-color: #e5ebff;
+    border-radius: 12.26px;
+    height: 24.52px;
+    margin-bottom: 10px;
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    cursor: pointer;
+}
+
+.hackathonBox .group2 {
+    height: 28.39px;
+    min-width: 102.66px;
+    position: relative;
+}
+
+.hackathonBox .textWrapper3 {
+    color: #000000cc;
+    font-size: 9.7px;
+    font-weight: 300;
+    letter-spacing: 0;
+    line-height: normal;
+    position: absolute;
+}
+
+.hackathonBox .overlapGroup {
+    border-radius: 5.16px;
+    height: 17px;
+    width: 48px;
+    position: relative;
+    top: 15px;
+    gap: 0.11rem;
+    display: flex;
+    align-items: center;
+    justify-content: start;
+}
+
+.hackathonBox .textWrapper4 {
+    color: var(--grey);
+    font-size: 8px;
+    font-weight: 400;
+    letter-spacing: 0;
+    line-height: normal;
+    padding: 0.3rem .5rem;
+    border: 1px solid var(--grey);
+    border-radius: 0.32263rem;
+    white-space: nowrap;
+}
+
+.hackathonBox .frame3 {
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+    gap: 12.91px;
+    position: relative;
+}
+
+.hackathonBox .frame4 {
+    align-items: flex-start;
+    display: flex;
+    gap: 12.09px;
+    height: 25.69px;
+    position: relative;
+    width: 150.35px;
+}
+
+.hackathonBox .textWrapperSmall {
+    color: var(--black);
+    font-size: .7rem;
+    font-weight: 600;
+}
+
+.hackathonBox .date,
+.mode {
+    height: 25.94px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    position: relative;
+    padding: .3rem;
+    border: 1px solid var(--blue);
+    border-radius: 0.5rem;
+}
+
+.buttonWrapper {
+    border: 0px none;
+}
+
+.hackathonBtn {
+    border-radius: .25rem;
+    padding: 0.3rem 1rem;
+    background-color: var(--blue);
+    color: var(--White);
+    border: 1px solid var(--blue);
+    font-size: .8rem;
+}
+
+.hackathonHeading {
+    color: var(--black);
+    display: flex;
+    align-items: center;
+    font-size: 2rem;
+    padding: 1rem;
+    width: 100%;
+}
