+++ src/modules/Dashboard/modules/Hackathon/pages/HackathonDetails.tsx
@@ -4,7 +4,7 @@ import { useNavigate, useParams } from "react-router-dom";
 import { getHackDetails } from "../services/HackathonApis";
 import { PowerfulButton } from "@/MuLearnComponents/MuButtons/MuButton";
 import { HackList } from "../services/HackathonInterfaces";
-import styles from "./Hackathon.module.css";
+import styles from "./HackathonCreate.module.css";
 
 type Props = {};
